moved tests to its own file and imported
